+++ applications/CSharpWrapperApplication/custom_includes/model_part_wrapper.h
@@ -1,12 +1,31 @@
+//    _____  _____ _                  __          __                                                 _ _           _   _
+//   / ____|/ ____| |                 \ \        / /                               /\               | (_)         | | (_)
+//  | |    | (___ | |__   __ _ _ __ _ _\ \  /\  / / __ __ _ _ __  _ __   ___ _ __ /  \   _ __  _ __ | |_  ___ __ _| |_ _  ___  _ __
+//  | |     \___ \| '_ \ / _` | '__| '_ \ \/  \/ / '__/ _` | '_ \| '_ \ / _ \ '__/ /\ \ | '_ \| '_ \| | |/ __/ _` | __| |/ _ \| '_  |
+//  | |____ ____) | | | | (_| | |  | |_) \  /\  /| | | (_| | |_) | |_) |  __/ | / ____ \| |_) | |_) | | | (_| (_| | |_| | (_) | | | |
+//   \_____|_____/|_| |_|\__,_|_|  | .__/ \/  \/ |_|  \__,_| .__/| .__/ \___|_|/_/    \_\ .__/| .__/|_|_|\___\__,_|\__|_|\___/|_| |_|
+//                                 | |                     | |   | |                    | |   | |
+//                                 |_|                     |_|   |_|                    |_|   |_|
+//
+//
+//  License: BSD License
+//   license: CSharpWrapperApplication/license.txt
+//
+//  Main authors:    Hubert Balcerzak
 
 #ifndef KRATOSMULTIPHYSICS_MODEL_PART_WRAPPER_H
 #define KRATOSMULTIPHYSICS_MODEL_PART_WRAPPER_H
 
+// System includes
+#include <vector>
+
+// External includes
+
+// Project includes
 #include "includes/model_part.h"
 #include "id_translator.h"
 #include "mesh_converter.h"
 #include "includes/variables.h"
-#include <vector>
 
 
 namespace CSharpKratosWrapper {
