+++ applications/CSharpWrapperApplication/custom_sources/kratos_internals.cpp
@@ -20,7 +20,7 @@
 // External includes
 
 // Project includes
-#include "kratos_internals.h"
+#include "custom_includes/kratos_internals.h"
 #include "utilities/variable_utils.h"
 #include "utilities/read_materials_utility.h"
 #include "custom_constitutive/hyper_elastic_isotropic_neo_hookean_3d.h"
